+++ src/main/java/auth/dto/request/TokenRequest.java
@@ -1,8 +1,11 @@
-package auth;
+package auth.dto.request;
 
 public class TokenRequest {
     private String username;
     private String password;
+
+    public TokenRequest() {
+    }
 
     public TokenRequest(String username, String password) {
         this.username = username;
