+++ src/main/java/auth/exception/AuthenticationException.java
@@ -1,4 +1,4 @@
-package auth;
+package auth.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
