+++ src/main/java/auth/presentation/LoginController.java
@@ -1,9 +1,14 @@
-package auth;
+package auth.presentation;
 
+import auth.dto.request.TokenRequest;
+import auth.dto.response.TokenResponse;
+import auth.service.LoginService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 
+@RequestMapping("/login/token")
 public class LoginController {
     private LoginService loginService;
 
@@ -11,7 +16,7 @@ public class LoginController {
         this.loginService = loginService;
     }
 
-    @PostMapping("/login/token")
+    @PostMapping
     public ResponseEntity<TokenResponse> loginToken(@RequestBody TokenRequest tokenRequest) {
         TokenResponse token = loginService.createToken(tokenRequest);
         return ResponseEntity.ok(token);
