+++ src/main/java/auth/presentation/LoginMember.java
@@ -1,4 +1,4 @@
-package auth;
+package auth.presentation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
