+++ src/main/java/auth/presentation/LoginMemberArgumentResolver.java
@@ -1,5 +1,7 @@
-package auth;
+package auth.presentation;
 
+import auth.exception.AuthenticationException;
+import auth.service.LoginService;
 import org.springframework.core.MethodParameter;
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.bind.support.WebDataBinderFactory;
