+++ src/main/java/nextstep/dto/request/ThemeRequest.java
@@ -7,6 +7,10 @@ public class ThemeRequest {
     private String desc;
     private int price;
 
+    public ThemeRequest() {
+
+    }
+
     public ThemeRequest(String name, String desc, int price) {
         this.name = name;
         this.desc = desc;
